requirements.txt
PyMuPDF==1.23.14
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langdetect==1.0.9
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
transformers==4.35.0
groq==0.4.1
python-dotenv==1.0.0
flask==2.3.3